Refactor search to be server rendered
I realised that the search results weren't
being server rendered. Before this commit, when the
user searched for something, the client side was
making an ajax request and modifying the DOM. I
extracted this functionality to the server side by
getting the search form to submit a get request,
take the paramaters of the querystring and use it to
make a call to the API and then render the index page
again with the new data.

NB I had to change a number of tests to reflect the
change and now that the client side isn't responsible
I have removed unnecerssary tests from cypress.