No results from search returns a message to user
Before this commit, when the user entered a search
that returned no results it was kicking an error
because the index ejs page was trying to loop through
the response to render the articles. I added a
conditional that checks if the response was undefined,
and if it is then return only a div that says the
search resulted in no matching results.